backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-tenants==3.5.0
psycopg2-binary==2.9.9
Pillow==10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
